requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
docopt==0.6.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
monotonic==1.6
openai==1.59.9
pyasyncore==1.0.4
pydantic==2.10.5
pydantic_core==2.27.2
pyinotify==0.9.6
python-valve==0.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
